requirements.txt
aioodbc==0.5.0
alembic==1.13.2
asyncpg==0.29.0
APScheduler==3.11.0
black==25.1.0
colorama==0.4.6
fastapi==0.115.0
fastapi-socketio==0.0.10
greenlet==3.0.3
gunicorn==21.2.0
httpx==0.27.0
isort==6.0.1
mypy==1.16.1
pytz==2025.2
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-multipart==0.0.9
python-socketio==5.11.4
PyJWT==2.8.0
SQLAlchemy==2.0.35
sqlmodel==0.0.22
uvicorn==0.30.1
